fix: Chromium browser error when hovering on tooltip items
In Chromium-based browsers, the event.relatedTarget property in the
`handleMouseLeave` function can sometimes be null or not an instance of Node,
leading to an error as below when using the `contains` method.

"TypeError: Failed to execute 'contains' on 'Node': parameter 1 is not of type 'Node'."

To solve this, updated the `handleMouseLeave` function to ensure that `event.relatedTarget`
is a valid Node before using it with `contains`.
1. Check if `event.relatedTarget` is a valid Node before using it with `contains`.
2. Update the condition in the `if` statement to include this check.